ahhhh remove tetsinput, check fracdynamic, yeah

+++ FractionalDynamic.java
@@ -1,14 +1,14 @@
-public class FractionalDynamic {
+public class FractionalDynamic{
     public static double knapsackFractionalDynamic(double[] weights, double[] values, int capacity) {
         int n = weights.length;
-        double[][] dp = new double[n + 1][(int) (capacity) + 1];
+        double[][] dp = new double[n + 1][capacity + 1];
 
         for (int i = 0; i <= n; i++) {
             for (int w = 0; w <= capacity; w++) {
                 if (i == 0 || w == 0) {
                     dp[i][w] = 0;
                 } else if (weights[i - 1] <= w) {
-                    dp[i][w] = Math.max(values[i - 1] + dp[i - 1][(int) (w - weights[i - 1])], dp[i - 1][w]);
+                    dp[i][w] = Math.max(values[i - 1] + dp[i - 1][w - (int) weights[i - 1]], dp[i - 1][w]);
                 } else {
                     dp[i][w] = dp[i - 1][w];
                 }
@@ -19,4 +19,5 @@ public class FractionalDynamic {
     }
 }
 
-//FIXME: TODO: AHHHHHH sTINKY: FIXME; 
+
+//TODO: FIXME: double check this tbh ig ;/
